pyproject.toml
[tool.poetry]
name = "postbox"
version = "0.1.0"
description = ""
authors = ["limsu1015 <tngur1015@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.10"
djangorestframework-simplejwt = "^5.3.1"
serializers = "^0.2.4"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
coverage = "^7.6.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.coverage.report]
fail_under = 80

[tool.coverage.run]
omit = [
  "*/migrations/*",
  "*/*test*.py",
]
